Poprawiono zapis rozmiaru czcionki
Była zapisywana w pikselach, teraz jest zapisywana w uniwersalnych metrach

+++ Editor/Text.pde
@@ -7,7 +7,7 @@ class Text
   public color col;
   public String content = new String();
   public float rotation = 0;
-  public int size = 32;
+  public float size = 32;
   
   Text(int x,int y,color col)
   {
